pyproject.toml
[project]
name = "vision-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.3",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
